pyproject.toml
[tool.poetry]
name = "spacetraders_sdk"
version = "2.3.1.9"
authors = ["Felix Baumkötter <stsdk@feba66.de>"]
description = "A python sdk for the SpaceTradersAPI game"
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
dacite = "^1.8.1"
python-dotenv = "^1.0.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
pytest-cov = "^4.1.0"
pydantic = "^2.3.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Homepage" = "https://github.com/feba66/spacetraders-sdk-v3"
"Bug Tracker" = "https://github.com/feba66/spacetraders-sdk-v3/issues"

[tool.autopep8]
max_line_length = 120
# in-place = true
# recursive = true
# aggressive = 3